feat: revise plotting functions to select runs
The commit adds a few lines of code in each plotting function to limit
the data used to a subset of the runs, selected according to whether the
probability of a certain policy reaches a value higher than 0.5. This allows to
discern between optimal and suboptimal agents. The policy is
selected with an additional argument in the visualization command parser.